server/projeto-clareza-backend/requirements.txt
amqp==5.3.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.4.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
